FEAT: Added Open close Garage Btn and Lock Unlock Garage Btn
The Door Page contains 2 buttons. One to open or close the garage door. The other to stop the door at a particular time. The stop is refered to as LOCK which the release is refered to as Unlock state. In case The door is locked, but open or close button is pressed the loc is removed from the server side and the open and close function is given priority.