FEAT: Fixed garage door status and Added Co sensor stat
The Co sensor API endpoint as well as the linearProgressBar backend was implemented. Based on the values defines in the config getCoColor() the color of the progress bar changes. The issue with the Garage door status not showing correct value was also fixed since the issue was only replated to the new API response change. Based on the Co level value, the LinearProgess changes fro the progressBar.