FEAT: Push Notification android and Door Animation with new connection status
Added the one signal dependency to the project to enable push notifications on android using FCM as the backbone. Added lottie animation for open ad closing animation of the door. The animation can be controlled by using animation controller. The new visual connection sttus infrom the user about the onnection with the server. If the connection is successful it turns green but it gets red in case the stus returned after the comand is not 1. Inititially the color of the connection status is blue.