FEAT: Added JWT token, Login system, Token Refresh, token baed door get route
The splashscreen check for the token locally, if not present redirect to login page. In case of login, the username and password are sent to the server and in case they are successful they are returned back with the jwt token. These data are sotred locally and can be accessed later on. In case of the token is present locally the slashscreen validate the token with the server. if the token is valid but expireed a new request to refresh the token is sent. In that case the new token is saved locally and can be used to access all the different routes. The door get routes uses the token to validate the requests. In case the token is not valid the stored username and passwrod is used to refresh the token. Some of the warning were also removed: from 91 to 41 only now.