FEAT: Added logout btn and fixed broken Login
Login system was not working due to missing Device parameter, not being added to the request body. A new logout btn was also added into the userpage to make the user exit the application if need be. Extra check for None value as the user id and password was added to make sure the refresh token understand that the user logged out.