FEAT: Added Guest Dash UI
New Role button was added into the loginScreen although the login screen is still missing the animation to be used for the role toggle switch. This new toggle switch, switches between the different login URI to connect the login system to the correct endpoint. Based the login status and the role, the HomeScreen or guestScreen are shown.