FEAT: Added Vehicle Panels, Add Vehicle btn and remove vehicle.
When user enter the User Page he will be greeted by a vehicles panels in which list of all the vehicles added linked to the owner will be present. Right now the UI for the information about the service data is still pending but will be added as soon as available. A custom dialog will appear if the user can to enter the data for new vehicle. The limit for max vehicle is still at 3. Also the validation fro the add vehicle dialog is still pending.